use baseName instead of paramName for the annotations
This fixes the analogon to [Issue #1219 of swagger-codegen](https://github.com/swagger-api/swagger-codegen/issues/1219), i.e. that the names of parameters get sanitized, and then the original names can't be matched anymore.

In addition I added some line breaks, so the annotations for a parameter are on lines before the parameter itself.